+++ renovate.json
@@ -1,12 +1,7 @@
 {
   "$schema": "https://docs.renovatebot.com/renovate-schema.json",
-  "extends": [
-    "github>angular/dev-infra//renovate-presets/default.json5"
-  ],
-  "baseBranchPatterns": [
-    "main",
-    "21.0.x"
-  ],
+  "extends": ["github>angular/dev-infra//renovate-presets/default.json5"],
+  "baseBranchPatterns": ["main", "21.0.x"],
   "ignoreDeps": [
     "@types/selenium-webdriver",
     "angular-1.5",
@@ -27,19 +22,11 @@
   ],
   "packageRules": [
     {
-      "matchFileNames": [
-        "modules/ssr-benchmarks/package.json",
-        "vscode-ng-language-service/integration/pre_standalone_project/package.json"
-      ],
+      "matchFileNames": ["modules/ssr-benchmarks/package.json"],
       "enabled": false
     },
     {
-      "matchManagers": [
-        "bazel",
-        "bazel-module",
-        "bazelisk",
-        "npm"
-      ],
+      "matchManagers": ["bazel", "bazel-module", "bazelisk", "npm"],
       "postUpgradeTasks": {
         "commands": [
           "git restore .npmrc",
@@ -60,12 +47,8 @@
     },
     {
       "enabled": false,
-      "matchDepNames": [
-        "tailwindcss"
-      ],
-      "matchUpdateTypes": [
-        "major"
-      ]
+      "matchDepNames": ["tailwindcss"],
+      "matchUpdateTypes": ["major"]
     }
   ]
 }
