--- vscode-ng-language-service/integration/pre_standalone_project/BUILD.bazel
@@ -1,12 +1,0 @@
-load("@aspect_bazel_lib//lib:copy_to_bin.bzl", "copy_to_bin")
-load("@npm//:defs.bzl", "npm_link_all_packages")
-
-npm_link_all_packages(name = "node_modules")
-
-copy_to_bin(
-    name = "pre_standalone_project",
-    srcs = glob(["**"]),
-    visibility = [
-        "//vscode-ng-language-service/integration:__pkg__",
-    ],
-)
