--- vscode-ng-language-service/integration/pre_standalone_project/app/app.component.ts
@@ -1,11 +1,0 @@
-import {Component, EventEmitter, Input, Output} from '@angular/core';
-
-@Component({
-  selector: 'my-app',
-  template: `<h1>Hello {{ name }}</h1>`,
-})
-export class AppComponent {
-  name = 'Angular';
-  @Input() appInput = '';
-  @Output() appOutput = new EventEmitter<string>();
-}
