--- vscode-ng-language-service/integration/pre_standalone_project/app/foo.component.html
@@ -1,4 +1,0 @@
-{{title | uppercase}}
-<span class="subtitle">
-    subtitle
-</span>
