--- vscode-ng-language-service/integration/pre_standalone_project/app/foo.component.ts
@@ -1,8 +1,0 @@
-import {Component} from '@angular/core';
-
-@Component({
-  templateUrl: 'foo.component.html',
-})
-export class FooComponent {
-  title = 'Foo Component';
-}
