--- vscode-ng-language-service/integration/pre_standalone_project/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "angular-ls-integration-test-project",
-  "private": true,
-  "dependencies": {
-    "@angular/common": "18.2.10",
-    "@angular/core": "18.2.10",
-    "zone.js": "0.15.0"
-  }
-}
