--- vscode-ng-language-service/integration/pre_standalone_project/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "moduleResolution": "node",
-    "experimentalDecorators": true,
-    "target": "es2015",
-    "strict": true,
-    "typeRoots": [
-      "node_modules/@types"
-    ]
-  },
-  "angularCompilerOptions": {
-    "strictTemplates": true,
-    "strictInjectionParameters": true
-  }
-}
