+++ vscode-ng-language-service/integration/project/package.json
@@ -2,10 +2,10 @@
   "name": "angular-ls-integration-test-project",
   "private": true,
   "dependencies": {
-    "@angular/common": "21.1.0-next.4",
-    "@angular/compiler": "21.1.0-next.4",
-    "@angular/compiler-cli": "21.1.0-next.4",
-    "@angular/core": "21.1.0-next.4",
+    "@angular/common": "workspace:*",
+    "@angular/compiler": "workspace:*",
+    "@angular/compiler-cli": "workspace:*",
+    "@angular/core": "workspace:*",
     "rxjs": "7.8.2"
   },
   "devDependencies": {
