+++ vscode-ng-language-service/integration/test_constants.ts
@@ -11,11 +11,6 @@ export function makeTempDir(): string {
 export const PACKAGE_ROOT = resolve(__dirname, '..');
 export const SERVER_PATH = join(PACKAGE_ROOT, 'server', 'index.js');
 export const PROJECT_PATH = join(PACKAGE_ROOT, 'integration', 'project');
-export const PRE_STANDALONE_PROJECT_PATH = join(
-  PACKAGE_ROOT,
-  'integration',
-  'pre_standalone_project',
-);
 
 export const APP_COMPONENT = join(PROJECT_PATH, 'app', 'app.component.ts');
 export const APP_COMPONENT_URI = pathToFileURL(APP_COMPONENT).href;
